Filling ratio + take bounding box as decimals
Non-finished code for the filling ratio characteristic.
When parsing the annotations, the bounding box corners should indicate
a position in the image as an integer, referring to a some pixel. It
could be changed again if we needed them to be floats in the future,
but decimals seems the best option for now.